backend/requirements.txt
aiogram==2.24
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.5
django-cors-headers==3.13.0
django-filter==22.1
django-parler==2.3
django-parler-rest==2.2
django-tabular-export
django-ckeditor==6.5.1
django-jazzmin==2.6.0
django-pgviews==0.5.7
django-jet-reboot==1.3.3
django-log-viewer==1.1.6
django-admin-sortable2==2.1.4
djangorestframework-gis~=1.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.25.1
rest-pandas==1.1.0
et-xmlfile==1.1.0
gunicorn==20.1.0
inflection==0.5.1
isort==5.11.4
jsonschema==4.17.3
kombu==5.2.4
matplotlib==3.7.1
numpy==1.24.1
openpyxl==3.0.10
pandas==1.5.2
Pillow==9.4.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
PyYAML==6.0
redis==4.4.0
six==1.16.0
scipy==1.10.1
sqlparse==0.4.3
uritemplate==4.1.1
uvicorn==0.20.0
vine==5.0.0
wcwidth==0.2.5
XlsxWriter==3.0.6
statsmodels==0.13.5
scikit-learn==1.2.2
phonenumbers==8.13.8